# contrib/openapi.yml
swagger: "2.0"
info:
  title: Kostyor API
  description: Aims & Launches Grenades
  version: 1.0.0
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /clusters:
    get:
      summary: List Clusters
      responses:
        200:
          description: An array of clusters.
          schema:
            type: array
            items:
              $ref: '#/definitions/Cluster'
  /clusters/{id}:
    get:
      summary: Show Cluster
      parameters:
        - name: id
          type: string
          pattern: &UUID_PATTERN |
            ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
          required: true
          in: path
          description: |
            Unique identifier representing the specific cluster.
      responses:
        200:
          description: A cluster.
          schema:
            $ref: '#/definitions/Cluster'
        404:
          description: Cluster not found.
          schema:
            $ref: '#/definitions/Error'
  /clusters/{id}/hosts:
    get:
      summary: List hosts in a cluster
      parameters:
        - name: id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific cluster.
      responses:
        200:
          description: List all hosts in the cluster.
          schema:
            type: array
            tems:
              $ref: '#/definitions/Host'
        404:
          description: Cluster not found.
          schema:
            $ref: '#/definitions/Error'
  /clusters/{id}/services:
    get:
      summary: List services in a cluster
      parameters:
        - name: id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific cluster.
      responses:
        200:
          description: List of all services in a cluster.
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
        404:
          description: Cluster not found.
          schema:
            $ref: '#/definitions/Error'
  /clusters/{id}/upgrades:
    get:
      summary: List Upgrade Tasks of Clusters
      parameters:
        - name: id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific cluster.
      responses:
        200:
          description: An arrays of upgrade tasks.
          schema:
            type: array
            items:
              $ref: '#/definitions/Upgrade'
        404:
          description: Cluster not found.
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Start Cluster Upgrade
      parameters:
        - name: id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific cluster.
      responses:
        202:
          description: A task instance to track the progress.
          schema:
            $ref: '#/definitions/Upgrade'
        400:
          description: Cluster is not ready for upgrade.
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Cluster not found.
          schema:
            $ref: '#/definitions/Error'
  /clusters/{id}/upgrades/{upgrade_id}:
    get:
      summary: Show Upgrade Task
      parameters:
        - name: id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific task.
      responses:
        200:
          description: A task.
          schema:
            $ref: '#/definitions/Upgrade'
        404:
          description: Task not found.
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Control Upgrade Flow of a Cluster
      parameters:
        - name: id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific cluster.
        - name: upgrade_id
          type: string
          pattern: *UUID_PATTERN
          required: true
          in: path
          description: |
            Unique identifier representing the specific upgrade task.
      responses:
        200:
          description: A task instance to track the progress.
          schema:
            $ref: '#/definitions/Upgrade'
        400:
          description: Cluster is not ready for upgrade.
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Either cluster or upgrade task not found.
          schema:
            $ref: '#/definitions/Error'
definitions:
  Cluster:
    type: object
    required:
      - id
      - name
      - version
      - status
    properties:
      id:
        type: string
        pattern: *UUID_PATTERN
      name:
        type: string
      version:
        type: string
        enum: &OPENSTACK_VERSIONS
          - unknown
          - liberty
          - mitaka
          - newton
      status:
        type: string
        enum: &ENVIRONMENT_STATUSES
          - READY_FOR_UPGRADE
          - UPGRADE_IN_PROGRESS
          - UPGRADE_PAUSED
          - UPGRADE_ERROR
          - UPGRADE_CANCELLED
          - NOT_READY_FOR_UPGRADE
          - ROLLBACK_IN_PROGRESS
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
        pattern: ^\w+$
      message:
        type: string
  Upgrade:
    type: object
    required:
      - id
      - status
      - cluster_id
    properties:
      id:
        type: string
        pattern: *UUID_PATTERN
      status:
        type: string
        enum:
          - pending
          - running
          - ready
          - error
      cluster_id:
        type: string
        description: |
          A unique identifier representing assigned cluster.
  Host:
    type: object
    required:
      - id
      - hostname
      - cluster_id
    properties:
      id:
        type: string
        pattern: *UUID_PATTERN
      hostname:
        type: string
      cluster_id:
        type: string
        pattern: *UUID_PATTERN
        description: |
          A unique identifier representing assigned cluster.
  Service:
    type: object
    required:
      - id
      - name
      - host_id
      - version
    properties:
      id:
        type: string
        pattern: *UUID_PATTERN
      name:
        type: string
      host_id:
        type: string
        pattern: *UUID_PATTERN
        description: |
          A unique identifier representing assigned host.
      version:
        type: string
        enum: *OPENSTACK_VERSIONS